(partially) address TOCTOU
When working with filesystems, especially, code that vaguely looks like
this:

```
if !a_file.exists() {
	do_something_to_create_file()
}
```

has a bug, of varying degrees of severity. This is called "Time-of-check
to time-of-use," and the problem is roughly:

```
if !a_file.exists() {
	// what happens if we get here, but before
	// do_something_to_create_file is called, we
	// swap to another thread, and another process creates
	// a_file, possibly by doing something fancy with symlinks
	do_something_to_create_file()
}
```

I don't think there is an actual problem here, probably, but it's better
just to "confidently" call these functions, rather than try and get
fancy. That way the creation is "atomic," if you will.

For more: https://en.wikipedia.org/wiki/Time-of-check_to_time-of-use

Now, the commit line says "partially" because

```
if !extract_destination.exists() {
    let _ = download_and_extract(
```

This is doing the same thing, but modifying all of that code to also be
okay with being called this way is a larger lift, and you don't have any
tests, so I am nervous about breaking something :)